backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
aiofiles
openai
nbformat
google-generativeai